add local transform data to avatar class, move Looker
